Hochob: Visitor Center Code Base
A visitor center or centre (see American and British English spelling
differences), visitor information center, tourist information center,
is a physical location that provides tourist information to visitors [0].

Code base is taken from the Bacab Home project [1]: Old god of the interior
of the earth and of thunder, sky-carrier, fourfold. Today? a Smart Home!
The documented journey to develop it ...

[0] https://en.wikipedia.org/wiki/Visitor_center
[1] https://github.com/bacabhome

Signed-off-by: Abraham Arce <xe1gyq@gmail.com>