Applications: Restricted Zone Notifier CPP
This application is designed to detect the humans present in a
predefined selected assembly line area. If the people enters the
marked assembly area, it raises the alert and sends through mqtt.
It is intended to demonstrate how to use CV to improve assembly
line safety for human operators and factory workers.

[0] https://github.com/intel-iot-devkit/restricted-zone-notifier-cpp

Signed-off-by: Abraham Arce <abraham.arce.moreno@intel.com>